payroll.Java
This payroll has been made in Java programming language, The user needs to input his/her data (name, position, hourly rate, worked hour) the hourly rate will be multiplied to worked-hour, and the result will be the Regular salary, after that, if the user worked greater than 9 (>9) the program will consider it as overtime, so, the Regular salary will be multiplied to two (2), and the result is the Overtime salary, and after that, the Overtime salary will be subtracted to insurance (it can be edited) and the result will be the Final Salary, and if the worked-hour of the user is below 8, the program will not consider it as overtime, so, the regular salary subtract to insurance and then the result is the Final salary. this program is just based on the Activity as a college student.